Code refactoring to configure gateway logs properly
