+++ services/gateway/app.js
@@ -17,15 +17,15 @@ const {
 	Microservice,
 	Logger,
 	LoggerConfig,
-	// Libs,
+	Libs,
 } = require('lisk-service-framework');
 
 const { MoleculerError } = require('moleculer').Errors;
 
 const SocketIOService = require('./shared/moleculer-io');
 
-// const ApiService = Libs['moleculer-web'];
-const ApiService = require('./shared/moleculer-web');
+const ApiService = Libs['moleculer-web'];
+const { methods } = require('./shared/moleculer-web/methods');
 
 const config = require('./config');
 const routes = require('./routes');
@@ -103,7 +103,7 @@ broker.createService({
 
 		logRequestParams: 'debug',
 		logResponseData: 'debug',
-        logRequest: 'debug',
+		logRequest: 'debug',
 		log2XXResponses: 'debug',
 		httpServerTimeout: 30 * 1000, // ms
 		optimizeOrder: true,
@@ -126,6 +126,7 @@ broker.createService({
 			namespaces,
 		},
 	},
+	methods,
 	events: {
 		'block.change': (payload) => sendSocketIoEvent('update.block', payload),
 		'round.change': (payload) => sendSocketIoEvent('update.round', payload),
