+++ services/gateway/shared/status.js
@@ -117,7 +117,7 @@ const getReady = () => {
 		lisk_blocks,
 		lisk_transactions,
 		lisk_accounts,
-		// lisk_delegates,
+		lisk_delegates,
 		// lisk_peers,
 	} = svcStatus;
 
@@ -126,7 +126,7 @@ const getReady = () => {
 			lisk_blocks,
 			lisk_transactions,
 			lisk_accounts,
-			// lisk_delegates,
+			lisk_delegates,
 			// lisk_peers,
 		},
 	};
@@ -149,7 +149,7 @@ const init = () => {
 	waitForIt(() => checkApiMapBoolean('/blocks', 'lisk_blocks'), 1500);
 	waitForIt(() => checkApiMapBoolean('/transactions', 'lisk_transactions'), 1500);
 	waitForIt(() => checkApiMapBoolean('/accounts', 'lisk_accounts'), 1500);
-	// waitForIt(() => checkApiMapBoolean('/delegates', 'lisk_delegates'), 1500);
+	waitForIt(() => checkApiMapBoolean('/delegates', 'lisk_delegates'), 1500);
 	// waitForIt(() => checkApiMapBoolean('/peers', 'lisk_peers'), 1500);
 };
